+++ lib/class/update.class.php
@@ -1146,6 +1146,7 @@ class Update
         $tables    = [ 'cache_object_count', 'cache_object_count_run' ];
         $collation = (AmpConfig::get('database_collation', 'utf8_unicode_ci'));
         $charset   = (AmpConfig::get('database_charset', 'utf8'));
+        $engine    = ($charset == 'utf8mb4') ? 'InnoDB' : 'MYISAM';
         foreach ($tables as $table) {
             $sql = "CREATE TABLE IF NOT EXISTS `" . $table . "` (" .
               "`object_id` int(11) unsigned NOT NULL," .
@@ -1154,7 +1155,7 @@ class Update
               "`threshold` int(11) unsigned NOT NULL DEFAULT '0'," .
               "`count_type` varchar(16) NOT NULL," .
               "PRIMARY KEY (`object_id`, `object_type`, `threshold`, `count_type`)" .
-              ") ENGINE=MyISAM DEFAULT CHARSET=$charset COLLATE=$collation;";
+              ") ENGINE=$engine DEFAULT CHARSET=$charset COLLATE=$collation;";
             $retval &= Dba::write($sql);
         }