+++ modules/catalog/beets/beets.catalog.php
@@ -74,11 +74,12 @@ class Catalog_beets extends Beets\Catalog
     {
         $collation = (AmpConfig::get('database_collation', 'utf8_unicode_ci'));
         $charset   = (AmpConfig::get('database_charset', 'utf8'));
+        $engine    = ($charset == 'utf8mb4') ? 'InnoDB' : 'MYISAM';
 
         $sql = "CREATE TABLE `catalog_beets` (`id` INT( 11 ) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY , " .
                 "`beetsdb` VARCHAR( 255 ) COLLATE $collation NOT NULL , " .
                 "`catalog_id` INT( 11 ) NOT NULL" .
-                ") ENGINE = MYISAM DEFAULT CHARSET=$charset COLLATE=$collation";
+                ") ENGINE = $engine DEFAULT CHARSET=$charset COLLATE=$collation";
         Dba::query($sql);
 
         return true;