+++ modules/catalog/seafile/seafile.catalog.php
@@ -103,6 +103,7 @@ class Catalog_Seafile extends Catalog
     {
         $collation = (AmpConfig::get('database_collation', 'utf8_unicode_ci'));
         $charset   = (AmpConfig::get('database_charset', 'utf8'));
+        $engine    = ($charset == 'utf8mb4') ? 'InnoDB' : 'MYISAM';
 
         $sql = "CREATE TABLE `" . self::$table_name . "` (" .
             "`id` INT UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY , " .
@@ -111,7 +112,7 @@ class Catalog_Seafile extends Catalog
             "`library_name` VARCHAR( 255 ) COLLATE $collation NOT NULL , " .
             "`api_call_delay` INT NOT NULL , " .
             "`catalog_id` INT( 11 ) NOT NULL" .
-            ") ENGINE = MYISAM DEFAULT CHARSET=$charset COLLATE=$collation";
+            ") ENGINE = $engine DEFAULT CHARSET=$charset COLLATE=$collation";
 
         Dba::query($sql);