+++ modules/catalog/subsonic/subsonic.catalog.php
@@ -88,13 +88,14 @@ class Catalog_subsonic extends Catalog
     {
         $collation = (AmpConfig::get('database_collation', 'utf8_unicode_ci'));
         $charset   = (AmpConfig::get('database_charset', 'utf8'));
+        $engine    = ($charset == 'utf8mb4') ? 'InnoDB' : 'MYISAM';
 
         $sql = "CREATE TABLE `catalog_subsonic` (`id` INT( 11 ) UNSIGNED NOT NULL AUTO_INCREMENT PRIMARY KEY , " .
             "`uri` VARCHAR( 255 ) COLLATE $collation NOT NULL , " .
             "`username` VARCHAR( 255 ) COLLATE $collation NOT NULL , " .
             "`password` VARCHAR( 255 ) COLLATE $collation NOT NULL , " .
             "`catalog_id` INT( 11 ) NOT NULL" .
-            ") ENGINE = MYISAM DEFAULT CHARSET=$charset COLLATE=$collation";
+            ") ENGINE = $engine DEFAULT CHARSET=$charset COLLATE=$collation";
         $db_results = Dba::query($sql);
 
         return true;